Silence the watchdog complains about 100% CPU usage by loopTask
If a task in RTOS does not yield, watchdog complains. These warnings
then get printed on console every approx 2 sec. When vTaskDelay was
added the warnings do not appear any more. Not sure why this was never
needed on the other ESPs, but it definitelly is needed for my ESP32C3,
so adding it in #if defined block. As a sideeffect maybe some inputs get
read with a slight delay.